Support metadata for multipart upload
